+++ smawg/default_rules.py
@@ -4,15 +4,15 @@ See https://github.com/expurple/smawg for more info about the project.
 """
 
 import smawg.exceptions as exc
-from smawg._plugin_interface import _GameState, _TurnStage
+from smawg._plugin_interface import GameState, _TurnStage
 
-__all__ = ["_Rules"]
+__all__ = ["Rules"]
 
 
-class _Rules:
+class Rules:
     """Rules that are used by `smawg.Game` by default."""
 
-    def __init__(self, game: _GameState) -> None:
+    def __init__(self, game: GameState) -> None:
         """Create an instance that will work on provided `game` instance."""
         self.game = game